style(region-selection): update selection UI to match new UI/UX in ContactList for swipe-to-reveal/dismiss
